DEV-1542: consolidate dialect-specific tests under tests/dialects/
- git mv tests/test_sqlite_{json_extract,udfs}.py into tests/dialects/
- new tests/dialects/conftest.py with shared helpers and orders_model fixture
- lift TestMultiDialectGeneration, TestLogAliasPreservation,
  TestStringHygieneDialectTranslation into tests/dialects/test_multi_dialect_generation.py
- lift TestMedianPercentilePerDialect, TestStatAggsPerDialect, TestTsqlDialect,
  TestSqliteJsonExtractInGenerator into tests/dialects/test_generator_delegation.py
  (SQLGenerator-surface tests covering the dialect dispatch path)
- delete TestDialects (single mysql smoke check, covered by
  TestMultiDialectGeneration::test_basic_query[mysql]) and TestDialectMapping
  (duplicate of tests/dialects/test_registry.py)
- drop the 8 sqlite spot-checks from tests/dialects/test_sqlite.py now
  superseded by the relocated comprehensive coverage; keep the 3
  module-level smoke checks that pin the import contract
- update CLAUDE.md:198 to point at slayer/sql/dialects/ and
  tests/dialects/test_multi_dialect_generation.py
