+++ tests/dialects/test_sqlite_udfs.py
