+++ tests/test_sql_generator.py
@@ -509,20 +509,6 @@ class TestBareColumnNames:
         assert "amount" in sql.lower()
 
 
-class TestDialects:
-    async def test_mysql_dialect(self, orders_model: SlayerModel) -> None:
-        gen = SQLGenerator(dialect="mysql")
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="*:count")],
-            time_dimensions=[
-                TimeDimension(dimension=ColumnRef(name="created_at"), granularity=TimeGranularity.MONTH),
-            ],
-        )
-        sql = await _generate(gen, query, orders_model)
-        assert "COUNT(*)" in sql  # Basic check — dialect-specific output
-
-
 class TestFields:
     async def test_arithmetic_field(self, generator: SQLGenerator, orders_model: SlayerModel) -> None:
         """Arithmetic field generates CTE + outer SELECT."""
@@ -1774,974 +1760,6 @@ class TestNestedFields:
         )
         sql = await _generate(generator, query, orders_model)
         assert "__agg" not in sql, f"__aggN__ placeholder leaked into SQL:\n{sql}"
-
-
-class TestDialectMapping:
-    """Test _dialect_for_type resolves all supported datasource types."""
-
-    @pytest.mark.parametrize(
-        "ds_type,expected",
-        [
-            ("postgres", "postgres"),
-            ("postgresql", "postgres"),
-            ("mysql", "mysql"),
-            ("mariadb", "mysql"),
-            ("clickhouse", "clickhouse"),
-            ("bigquery", "bigquery"),
-            ("snowflake", "snowflake"),
-            ("sqlite", "sqlite"),
-            ("duckdb", "duckdb"),
-            ("redshift", "redshift"),
-            ("trino", "trino"),
-            ("presto", "presto"),
-            ("athena", "presto"),
-            ("databricks", "databricks"),
-            ("spark", "spark"),
-            ("mssql", "tsql"),
-            ("sqlserver", "tsql"),
-            ("tsql", "tsql"),
-            ("oracle", "oracle"),
-            (None, "postgres"),
-            ("unknown", "postgres"),
-        ],
-    )
-    def test_dialect_for_type(self, ds_type: str, expected: str) -> None:
-        assert SlayerQueryEngine._dialect_for_type(ds_type) == expected
-
-
-class TestMultiDialectGeneration:
-    """Test SQL generation across all supported dialects."""
-
-    @pytest.fixture
-    def orders_model(self) -> SlayerModel:
-        model = SlayerModel(
-            name="orders",
-            sql_table="public.orders",
-            data_source="test",
-            default_time_dimension="created_at",
-            columns=[
-                Column(name="id", sql="id", type=DataType.DOUBLE, primary_key=True),
-                Column(name="status", sql="status", type=DataType.TEXT),
-                Column(name="created_at", sql="created_at", type=DataType.TIMESTAMP),
-
-                Column(name="revenue", sql="amount", type=DataType.DOUBLE),
-                # Second numeric column so 2-arg stat aggregates
-                # (corr(other=...) / covar_*(other=...)) have a valid LHS+RHS pair.
-                Column(name="quantity", sql="quantity", type=DataType.DOUBLE),
-            ],
-        )
-        return model
-
-    ALL_DIALECTS = [
-        "postgres",
-        "mysql",
-        "sqlite",
-        "clickhouse",
-        "bigquery",
-        "snowflake",
-        "duckdb",
-        "redshift",
-        "trino",
-        "presto",
-        "databricks",
-        "spark",
-        "tsql",
-        "oracle",
-    ]
-
-    @pytest.mark.parametrize("dialect", ALL_DIALECTS)
-    async def test_basic_query(self, dialect: str, orders_model: SlayerModel) -> None:
-        """Basic aggregation query should generate valid SQL for every dialect."""
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="*:count"), ModelMeasure(formula="revenue:sum")],
-            dimensions=[ColumnRef(name="status")],
-        )
-        sql = await _generate(gen, query, orders_model)
-        assert "COUNT(" in sql
-        assert "SUM(" in sql
-
-    @pytest.mark.parametrize("dialect", ALL_DIALECTS)
-    async def test_date_trunc(self, dialect: str, orders_model: SlayerModel) -> None:
-        """DATE_TRUNC should produce valid output for every dialect."""
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="*:count")],
-            time_dimensions=[TimeDimension(dimension=ColumnRef(name="created_at"), granularity=TimeGranularity.MONTH)],
-        )
-        sql = await _generate(gen, query, orders_model)
-        assert "COUNT(" in sql
-        # Each dialect uses its own truncation function
-        sql_upper = sql.upper()
-        assert any(fn in sql_upper for fn in ["DATE_TRUNC", "STRFTIME", "TRUNC", "STR_TO_DATE", "DATETRUNC"])
-
-    @pytest.mark.parametrize("dialect", ["postgres", "mysql", "bigquery", "duckdb", "snowflake", "tsql"])
-    async def test_date_trunc_casts_unknown_typed_time_dim(self, dialect: str) -> None:
-        """A time-dimension whose ``sql`` is a bare literal (or any expression
-        whose live type is ``unknown``) must be wrapped in ``CAST(... AS
-        TIMESTAMP)`` before being passed to ``DATE_TRUNC``. Postgres has
-        multiple overloads keyed on the second argument's type and rejects
-        ``DATE_TRUNC('month', '2025-12-01')`` with ``AmbiguousFunctionError``.
-
-        Bare column references stay unwrapped — their live DB type is known,
-        and forcing a cast could strip ``TIMESTAMPTZ`` to ``TIMESTAMP``.
-        """
-        gen = SQLGenerator(dialect=dialect)
-        model = SlayerModel(
-            name="orders",
-            sql_table="public.orders",
-            data_source="test",
-            columns=[
-                Column(name="id", sql="id", type=DataType.INT, primary_key=True),
-                Column(name="ts", sql="'2025-12-01'", type=DataType.TIMESTAMP),
-                Column(name="created_at", sql="created_at", type=DataType.TIMESTAMP),
-            ],
-        )
-        # Bare-literal time dim — must be cast.
-        sql = await _generate(
-            gen,
-            SlayerQuery(
-                source_model="orders",
-                measures=[ModelMeasure(formula="*:count")],
-                time_dimensions=[TimeDimension(dimension=ColumnRef(name="ts"), granularity=TimeGranularity.MONTH)],
-            ),
-            model,
-        )
-        # sqlglot transpiles ``TIMESTAMP`` → ``DATETIME`` on MySQL / BigQuery,
-        # so we don't assert the literal target-type spelling — only that the
-        # literal is wrapped in a CAST.
-        assert "CAST('2025-12-01' AS" in sql, sql
-        # Bare-column time dim — must NOT be cast.
-        sql = await _generate(
-            gen,
-            SlayerQuery(
-                source_model="orders",
-                measures=[ModelMeasure(formula="*:count")],
-                time_dimensions=[TimeDimension(dimension=ColumnRef(name="created_at"), granularity=TimeGranularity.MONTH)],
-            ),
-            model,
-        )
-        assert "CAST(" not in sql.upper(), sql
-
-    @pytest.mark.parametrize("dialect", ALL_DIALECTS)
-    async def test_calendar_time_shift(self, dialect: str, orders_model: SlayerModel) -> None:
-        """Calendar-based time_shift should produce dialect-appropriate date arithmetic in shifted CTE."""
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            time_dimensions=[TimeDimension(dimension=ColumnRef(name="created_at"), granularity=TimeGranularity.MONTH)],
-            measures=[ModelMeasure(formula="revenue:sum"), ModelMeasure(formula="time_shift(revenue:sum, -1, 'year')", name="rev_prev_year")],
-        )
-        sql = await _generate(gen, query, orders_model)
-        assert "shifted_" in sql
-        assert "LEFT JOIN" in sql
-        # Join should be simple equality (timestamp shift is inside the shifted CTE)
-        # Dialect-specific date arithmetic should appear in the shifted CTE's SELECT/GROUP BY
-        sql_upper = sql.upper()
-        if dialect == "sqlite":
-            assert "DATE(" in sql_upper
-        elif dialect == "tsql":
-            assert "DATEADD" in sql_upper
-        else:
-            assert "INTERVAL" in sql_upper
-
-    @pytest.mark.parametrize("dialect", ["mysql", "clickhouse"])
-    async def test_window_measure_multi_unit_interval_dialect_correct(
-        self, dialect: str, orders_model: SlayerModel,
-    ) -> None:
-        """Multi-unit windows (e.g. '1y2m3d') must render as separate per-unit
-        INTERVAL clauses on MySQL and ClickHouse — never as a single
-        Postgres-shaped quoted multi-unit literal which neither dialect parses.
-
-        Codex flagged this as a real correctness bug during PR #64 review:
-        `_duration_interval_sql` had only two branches (SQLite + "Postgres-style"),
-        and the latter emitted `INTERVAL '1 year 2 month 3 day'` for every
-        non-SQLite dialect.
-        """
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            time_dimensions=[
-                TimeDimension(dimension=ColumnRef(name="created_at"),
-                              granularity=TimeGranularity.DAY),
-            ],
-            measures=[ModelMeasure(formula="revenue:sum(window='1y2m3d')",
-                                   name="rev_w")],
-        )
-        sql = await _generate(generator=gen, query=query, model=orders_model)
-        norm = _norm(sql).upper()
-        # The broken Postgres-shape multi-unit literal must NOT appear.
-        assert "INTERVAL '1 YEAR 2 MONTH 3 DAY'" not in norm, (
-            f"Multi-unit Postgres-shape INTERVAL literal is invalid on {dialect}.\n"
-            f"sql:\n{sql}"
-        )
-        # Per-unit INTERVAL clauses must each be present (sqlglot transpiles
-        # exp.Interval per dialect; MySQL + ClickHouse both render as
-        # `INTERVAL N UNIT` for these AST nodes).
-        for piece in ("INTERVAL 1 YEAR", "INTERVAL 2 MONTH", "INTERVAL 3 DAY"):
-            assert piece in norm, (
-                f"Expected dialect-correct '{piece}' in {dialect} output.\n"
-                f"sql:\n{sql}"
-            )
-
-    @pytest.mark.parametrize("dialect", ["mysql", "clickhouse"])
-    async def test_window_measure_single_unit_interval_dialect_correct(
-        self, dialect: str, orders_model: SlayerModel,
-    ) -> None:
-        """Even single-unit windows must render unquoted on MySQL/ClickHouse.
-
-        The pre-refactor code emits `INTERVAL '7 day'` for single-unit windows
-        on every non-SQLite dialect, which is invalid MySQL syntax (MySQL wants
-        `INTERVAL 7 DAY`). After the AST refactor, sqlglot's per-dialect
-        transpiler emits the canonical form for each dialect.
-        """
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            time_dimensions=[
-                TimeDimension(dimension=ColumnRef(name="created_at"),
-                              granularity=TimeGranularity.DAY),
-            ],
-            measures=[ModelMeasure(formula="revenue:sum(window='7d')",
-                                   name="rev_w")],
-        )
-        sql = await _generate(generator=gen, query=query, model=orders_model)
-        norm = _norm(sql).upper()
-        assert "INTERVAL '7 DAY'" not in norm, (
-            f"Quoted single-unit INTERVAL literal is invalid on {dialect}.\n"
-            f"sql:\n{sql}"
-        )
-        assert "INTERVAL 7 DAY" in norm, (
-            f"Expected dialect-correct 'INTERVAL 7 DAY' in {dialect} output.\n"
-            f"sql:\n{sql}"
-        )
-
-    async def test_window_measure_tsql_uses_dateadd(
-        self, orders_model: SlayerModel,
-    ) -> None:
-        """Window boundary on T-SQL must use DATEADD instead of INTERVAL (invalid T-SQL)."""
-        gen = SQLGenerator(dialect="tsql")
-        query = SlayerQuery(
-            source_model="orders",
-            time_dimensions=[
-                TimeDimension(dimension=ColumnRef(name="created_at"),
-                              granularity=TimeGranularity.DAY),
-            ],
-            measures=[ModelMeasure(formula="revenue:sum(window='7d')", name="rev_w")],
-        )
-        sql = await _generate(generator=gen, query=query, model=orders_model)
-        norm = _norm(sql).upper()
-        assert "DATEADD" in norm, f"Expected DATEADD in T-SQL window output:\n{sql}"
-        assert "INTERVAL" not in norm, f"INTERVAL is invalid T-SQL syntax:\n{sql}"
-
-    async def test_window_measure_tsql_multi_unit_uses_chained_dateadd(
-        self, orders_model: SlayerModel,
-    ) -> None:
-        """Multi-unit window '1y2m3d' on T-SQL must use chained DATEADD calls, not INTERVAL."""
-        gen = SQLGenerator(dialect="tsql")
-        query = SlayerQuery(
-            source_model="orders",
-            time_dimensions=[
-                TimeDimension(dimension=ColumnRef(name="created_at"),
-                              granularity=TimeGranularity.DAY),
-            ],
-            measures=[ModelMeasure(formula="revenue:sum(window='1y2m3d')", name="rev_w")],
-        )
-        sql = await _generate(generator=gen, query=query, model=orders_model)
-        norm = _norm(sql).upper()
-        # Must use DATEADD, not INTERVAL literals
-        assert "INTERVAL" not in norm, f"INTERVAL is invalid T-SQL syntax:\n{sql}"
-        assert "DATEADD" in norm, f"Expected DATEADD in multi-unit T-SQL window:\n{sql}"
-
-    # DEV-1317: cross-dialect stat-agg generation. The exact SQL shape per
-    # Tier-1 dialect is pinned in TestStatAggsPerDialect; here we just confirm
-    # the generator produces parseable SQL on every supported dialect.
-    # Assertions check the function-call SHAPE (qualified column refs in the
-    # arg slot), not just substring fragments — substrings like "STDDEV" or
-    # "CORR" pass even when the aggregate has regressed because aliases
-    # such as `revenue_stddev_samp` always contain the family name.
-
-    @pytest.mark.parametrize("dialect", ALL_DIALECTS)
-    @pytest.mark.parametrize(
-        "formula",
-        [
-            "revenue:stddev_samp",
-            "revenue:stddev_pop",
-            "revenue:var_samp",
-            "revenue:var_pop",
-        ],
-    )
-    async def test_one_arg_stat_agg_generation(
-        self,
-        dialect: str,
-        formula: str,
-        orders_model: SlayerModel,
-    ) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula=formula)],
-        )
-        sql = await _generate(generator=gen, query=query, model=orders_model)
-        upper = sql.upper()
-        assert "SELECT" in upper
-        # The aggregate must wrap the resolved value column (orders.amount,
-        # since the `revenue` Column has sql="amount") in its single-arg slot.
-        assert "(ORDERS.AMOUNT)" in upper, (
-            f"expected single-arg call (ORDERS.AMOUNT) in SQL for {formula!r} on {dialect}:\n{sql}"
-        )
-
-    # corr / covar_samp / covar_pop are implemented via variance-decomposition
-    # formula on MySQL and T-SQL (neither has native two-arg functions), so those
-    # dialects are filtered out of the direct two-arg call assertion matrix.
-    @pytest.mark.parametrize(
-        "dialect", [d for d in ALL_DIALECTS if d not in ("mysql", "tsql")],
-    )
-    @pytest.mark.parametrize(
-        "formula",
-        [
-            "revenue:corr(other=quantity)",
-            "revenue:covar_samp(other=quantity)",
-            "revenue:covar_pop(other=quantity)",
-        ],
-    )
-    async def test_two_arg_stat_agg_generation(
-        self,
-        dialect: str,
-        formula: str,
-        orders_model: SlayerModel,
-    ) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula=formula)],
-        )
-        sql = await _generate(generator=gen, query=query, model=orders_model)
-        upper = sql.upper()
-        assert "SELECT" in upper
-        # Both legs (LHS column AND `other=` kwarg) must be qualified and
-        # appear in the function-call's two-arg slot in that order. This
-        # asymmetry vs the 1-arg test is what distinguishes the test bodies
-        # for Sonar python:S4144 and pins the new `_resolve_agg_param` +
-        # `_resolve_value_sql` qualification path for 2-arg stats.
-        assert "(ORDERS.AMOUNT, ORDERS.QUANTITY)" in upper, (
-            f"expected two-arg call (ORDERS.AMOUNT, ORDERS.QUANTITY) in SQL for {formula!r} "
-            f"on {dialect}:\n{sql}"
-        )
-
-    @pytest.mark.parametrize("dialect", ["mysql", "tsql"])
-    @pytest.mark.parametrize(
-        "formula",
-        [
-            "revenue:corr(other=quantity)",
-            "revenue:covar_samp(other=quantity)",
-            "revenue:covar_pop(other=quantity)",
-        ],
-    )
-    async def test_two_arg_stat_formula_dialects_generate_valid_sql(
-        self,
-        dialect: str,
-        formula: str,
-        orders_model: SlayerModel,
-    ) -> None:
-        """MySQL and T-SQL emit variance-decomposition formula instead of direct two-arg call."""
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula=formula)],
-        )
-        sql = await _generate(generator=gen, query=query, model=orders_model)
-        assert "SELECT" in sql.upper()
-        # The formula uses division (variance decomposition)
-        assert "/" in sql
-
-    @pytest.mark.parametrize(
-        "formula", [
-            "revenue:corr(other=quantity)",
-            "revenue:covar_samp(other=quantity)",
-            "revenue:covar_pop(other=quantity)",
-        ],
-    )
-    async def test_two_arg_stat_agg_mysql_emits_formula_valid_sql(
-        self, formula: str, orders_model: SlayerModel,
-    ) -> None:
-        """MySQL uses variance-decomposition formula for corr/covar_samp/covar_pop."""
-        gen = SQLGenerator(dialect="mysql")
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula=formula)],
-        )
-        sql = await _generate(generator=gen, query=query, model=orders_model)
-        assert "SELECT" in sql.upper()
-        # Formula uses division (variance decomposition)
-        assert "/" in sql
-
-
-class TestSqliteJsonExtractInGenerator:
-    """DEV-1331: ``json_extract(col, '$.path')`` in ``Column.sql`` must not be
-    rewritten to ``col -> '$.path'`` on SQLite — the operator returns the
-    JSON-quoted form, silently breaking equality / CASE WHEN matches.
-    """
-
-    @pytest.fixture
-    def model_with_json_dim(self) -> SlayerModel:
-        return SlayerModel(
-            name="users",
-            sql_table="users",
-            data_source="test",
-            columns=[
-                Column(name="id", sql="id", type=DataType.DOUBLE, primary_key=True),
-                Column(name="payload", sql="payload", type=DataType.TEXT),
-                Column(
-                    name="tier",
-                    sql="json_extract(payload, '$.tier')",
-                    type=DataType.TEXT,
-                ),
-                Column(
-                    name="is_gold",
-                    sql=(
-                        "CASE LOWER(json_extract(payload, '$.tier')) "
-                        "WHEN 'gold' THEN 1 ELSE 0 END"
-                    ),
-                    type=DataType.DOUBLE,
-                ),
-            ],
-        )
-
-    async def test_sqlite_column_sql_with_json_extract_dimension(
-        self, model_with_json_dim: SlayerModel,
-    ) -> None:
-        gen = SQLGenerator(dialect="sqlite")
-        query = SlayerQuery(
-            source_model="users",
-            dimensions=[ColumnRef(name="tier")],
-            measures=[ModelMeasure(formula="*:count")],
-        )
-        sql = await _generate(generator=gen, query=query, model=model_with_json_dim)
-        assert "JSON_EXTRACT(" in sql, f"missing JSON_EXTRACT in:\n{sql}"
-        # The lossy ``payload -> '$.tier'`` form must not appear.
-        assert "payload -> '$.tier'" not in sql, sql
-
-    async def test_sqlite_column_sql_with_json_extract_in_case_when(
-        self, model_with_json_dim: SlayerModel,
-    ) -> None:
-        gen = SQLGenerator(dialect="sqlite")
-        query = SlayerQuery(
-            source_model="users",
-            measures=[ModelMeasure(formula="is_gold:sum")],
-        )
-        sql = await _generate(generator=gen, query=query, model=model_with_json_dim)
-        assert "JSON_EXTRACT(" in sql, sql
-        assert "payload -> '$.tier'" not in sql, sql
-
-    async def test_sqlite_inline_sql_subquery_with_json_extract(self) -> None:
-        model = SlayerModel(
-            name="users",
-            sql=(
-                "SELECT id, json_extract(payload, '$.tier') AS tier "
-                "FROM raw_users"
-            ),
-            data_source="test",
-            columns=[
-                Column(name="id", sql="id", type=DataType.DOUBLE, primary_key=True),
-                Column(name="tier", sql="tier", type=DataType.TEXT),
-            ],
-        )
-        gen = SQLGenerator(dialect="sqlite")
-        query = SlayerQuery(
-            source_model="users",
-            dimensions=[ColumnRef(name="tier")],
-            measures=[ModelMeasure(formula="*:count")],
-        )
-        sql = await _generate(generator=gen, query=query, model=model)
-        assert "JSON_EXTRACT(" in sql, sql
-        assert "payload -> '$.tier'" not in sql, sql
-
-    async def test_postgres_column_sql_with_json_extract_unchanged(
-        self, model_with_json_dim: SlayerModel,
-    ) -> None:
-        """Regression guard: rewrite is SQLite-only; Postgres path is untouched.
-
-        Postgres has no scalar-vs-JSON quoting bug for ``json_extract``;
-        sqlglot transpiles it to ``JSON_EXTRACT_PATH(j, 'k')``. We just
-        assert the generator produces *some* form of JSON extraction and
-        does not crash.
-        """
-        gen = SQLGenerator(dialect="postgres")
-        query = SlayerQuery(
-            source_model="users",
-            dimensions=[ColumnRef(name="tier")],
-            measures=[ModelMeasure(formula="*:count")],
-        )
-        sql = await _generate(generator=gen, query=query, model=model_with_json_dim)
-        assert "JSON_EXTRACT" in sql.upper(), sql
-
-
-class TestMedianPercentilePerDialect:
-    """Per-dialect SQL emission for median and percentile aggregations.
-
-    These pin the dialect-specific output of `_build_median` and
-    `_build_percentile` and assert that MySQL raises ``NotImplementedError``
-    (no native function, no Python-UDF mechanism).
-    """
-
-    def _measure(
-        self,
-        *,
-        agg: str,
-        agg_kwargs: dict[str, str] | None = None,
-    ) -> EnrichedMeasure:
-        return EnrichedMeasure(
-            name="amount",
-            sql="amount",
-            model_name="orders",
-            alias=f"amount_{agg}",
-            aggregation=agg,
-            agg_kwargs=agg_kwargs or {},
-        )
-
-    # --- median ------------------------------------------------------------
-
-    def test_build_median_postgres(self) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        inner = sqlglot.parse_one("amount", dialect="postgres")
-        sql = gen._build_median(inner).sql(dialect="postgres")
-        assert sql == "PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY amount)"
-
-    def test_build_median_sqlite_uses_udf_call(self) -> None:
-        gen = SQLGenerator(dialect="sqlite")
-        inner = sqlglot.parse_one("amount", dialect="sqlite")
-        sql = gen._build_median(inner).sql(dialect="sqlite")
-        # sqlglot rewrites MEDIAN(x) to PERCENTILE_CONT(x, 0.5) for SQLite,
-        # which our percentile_cont UDF handles. SQLite UDF lookup is
-        # case-insensitive.
-        assert sql == "PERCENTILE_CONT(amount, 0.5)"
-
-    def test_build_median_clickhouse_unchanged(self) -> None:
-        gen = SQLGenerator(dialect="clickhouse")
-        inner = sqlglot.parse_one("amount", dialect="clickhouse")
-        sql = gen._build_median(inner).sql(dialect="clickhouse")
-        # ClickHouse has native median(); sqlglot transpiles to its parametric form.
-        assert sql == "quantile(0.5)(amount)"
-
-    def test_build_median_duckdb(self) -> None:
-        gen = SQLGenerator(dialect="duckdb")
-        inner = sqlglot.parse_one("amount", dialect="duckdb")
-        sql = gen._build_median(inner).sql(dialect="duckdb")
-        # sqlglot translates PERCENTILE_CONT to DuckDB's QUANTILE_CONT.
-        assert "QUANTILE_CONT" in sql or "PERCENTILE_CONT" in sql
-
-    def test_build_median_mysql_raises(self) -> None:
-        gen = SQLGenerator(dialect="mysql")
-        inner = sqlglot.parse_one("amount", dialect="mysql")
-        with pytest.raises(NotImplementedError, match="MySQL"):
-            gen._build_median(inner)
-
-    # --- percentile --------------------------------------------------------
-
-    def test_build_percentile_postgres(self) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = self._measure(agg="percentile", agg_kwargs={"p": "0.95"})
-        sql = gen._build_percentile(m).sql(dialect="postgres")
-        assert sql == "PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY orders.amount)"
-
-    def test_build_percentile_sqlite(self) -> None:
-        gen = SQLGenerator(dialect="sqlite")
-        m = self._measure(agg="percentile", agg_kwargs={"p": "0.5"})
-        sql = gen._build_percentile(m).sql(dialect="sqlite")
-        assert sql == "PERCENTILE_CONT(orders.amount, 0.5)"
-
-    def test_build_percentile_clickhouse_emits_quantile(self) -> None:
-        gen = SQLGenerator(dialect="clickhouse")
-        m = self._measure(agg="percentile", agg_kwargs={"p": "0.75"})
-        sql = gen._build_percentile(m).sql(dialect="clickhouse")
-        # ClickHouse parametric aggregate syntax.
-        assert sql == "quantile(0.75)(orders.amount)"
-
-    @pytest.mark.parametrize("p", ["0.05", "0.25", "0.5", "0.95"])
-    def test_build_percentile_clickhouse_param_substitution(self, p: str) -> None:
-        gen = SQLGenerator(dialect="clickhouse")
-        m = self._measure(agg="percentile", agg_kwargs={"p": p})
-        sql = gen._build_percentile(m).sql(dialect="clickhouse")
-        assert sql == f"quantile({p})(orders.amount)"
-
-    def test_build_percentile_duckdb(self) -> None:
-        gen = SQLGenerator(dialect="duckdb")
-        m = self._measure(agg="percentile", agg_kwargs={"p": "0.5"})
-        sql = gen._build_percentile(m).sql(dialect="duckdb")
-        # sqlglot rewrites the WITHIN GROUP form to DuckDB's QUANTILE_CONT.
-        assert "QUANTILE_CONT" in sql
-        # Qualified column.
-        assert "orders.amount" in sql
-
-    def test_build_percentile_mysql_raises(self) -> None:
-        gen = SQLGenerator(dialect="mysql")
-        m = self._measure(agg="percentile", agg_kwargs={"p": "0.5"})
-        with pytest.raises(NotImplementedError, match="MySQL"):
-            gen._build_percentile(m)
-
-    def test_build_percentile_missing_p_raises(self) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = self._measure(agg="percentile", agg_kwargs={})
-        with pytest.raises(ValueError, match="requires parameter 'p'"):
-            gen._build_percentile(m)
-
-    def test_build_percentile_unsafe_p_rejected(self) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = self._measure(agg="percentile", agg_kwargs={"p": "0.5); DROP TABLE x; --"})
-        with pytest.raises(ValueError, match="Unsafe value"):
-            gen._build_percentile(m)
-
-    def test_build_percentile_uses_model_level_default_p(self) -> None:
-        """Model-level Aggregation(name='percentile', params=[p=...]) supplies the default."""
-        gen = SQLGenerator(dialect="postgres")
-        agg_def = Aggregation(
-            name="percentile",
-            params=[AggregationParam(name="p", sql="0.9")],
-        )
-        m = EnrichedMeasure(
-            name="amount",
-            sql="amount",
-            model_name="orders",
-            alias="amount_percentile",
-            aggregation="percentile",
-            agg_kwargs={},
-            aggregation_def=agg_def,
-        )
-        sql = gen._build_percentile(m).sql(dialect="postgres")
-        assert sql == "PERCENTILE_CONT(0.9) WITHIN GROUP (ORDER BY orders.amount)"
-
-    def test_build_percentile_query_kwarg_overrides_model_default(self) -> None:
-        """Query-time agg_kwargs win over the model-level default."""
-        gen = SQLGenerator(dialect="postgres")
-        agg_def = Aggregation(
-            name="percentile",
-            params=[AggregationParam(name="p", sql="0.9")],
-        )
-        m = EnrichedMeasure(
-            name="amount",
-            sql="amount",
-            model_name="orders",
-            alias="amount_percentile",
-            aggregation="percentile",
-            agg_kwargs={"p": "0.25"},
-            aggregation_def=agg_def,
-        )
-        sql = gen._build_percentile(m).sql(dialect="postgres")
-        assert sql == "PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY orders.amount)"
-
-    # --- A2: percentile p must be a numeric literal in [0, 1] -----------
-
-    def test_build_percentile_rejects_non_literal_p(self) -> None:
-        """`measure:percentile(p=quantity)` must fail at SQL-generation time
-        with a clear validation error, not silently emit a column reference
-        in PERCENTILE_CONT(p)'s direct-arg slot. Without this guard a
-        non-literal `p` flows through `_resolve_agg_param` (which is
-        identifier-friendly for the column-ref kwargs like `other=`), gets
-        rendered as `orders.quantity`, and fails at the database with a
-        dialect-specific error.
-        """
-        gen = SQLGenerator(dialect="postgres")
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias="amount_percentile", aggregation="percentile",
-            agg_kwargs={"p": "quantity"},
-        )
-        with pytest.raises(ValueError, match="numeric literal"):
-            gen._build_percentile(m)
-
-    def test_build_percentile_rejects_p_out_of_range(self) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias="amount_percentile", aggregation="percentile",
-            agg_kwargs={"p": "1.5"},
-        )
-        with pytest.raises(ValueError, match=r"\[0, 1\]"):
-            gen._build_percentile(m)
-
-    def test_build_percentile_rejects_p_negative(self) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias="amount_percentile", aggregation="percentile",
-            agg_kwargs={"p": "-0.1"},
-        )
-        with pytest.raises(ValueError, match=r"\[0, 1\]"):
-            gen._build_percentile(m)
-
-    def test_build_percentile_rejects_non_literal_p_via_model_default(self) -> None:
-        """Model-level defaults bypass `_validate_agg_param_value` (trust
-        model: model authors are trusted). The new numeric-literal check
-        catches anything that's not a number even on that path — closes
-        the gap where a malicious model author could put `p=pg_sleep(10)`
-        as a default. Codex review #3 on PR #82.
-        """
-        gen = SQLGenerator(dialect="postgres")
-        agg_def = Aggregation(
-            name="percentile",
-            params=[AggregationParam(name="p", sql="pg_sleep(10)")],
-        )
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias="amount_percentile", aggregation="percentile",
-            agg_kwargs={}, aggregation_def=agg_def,
-        )
-        with pytest.raises(ValueError, match="numeric literal"):
-            gen._build_percentile(m)
-
-
-class TestStatAggsPerDialect:
-    """Per-dialect SQL emission for the new statistical aggregations
-    (DEV-1317): stddev_samp, stddev_pop, var_samp, var_pop, corr,
-    covar_samp, covar_pop.
-
-    These pin the observed SQL output for each dialect — including
-    sqlglot's transpilation quirks (e.g., var_samp → VARIANCE on SQLite,
-    var_pop → VARIANCE_POP on SQLite/MySQL) — so the SQLite UDF
-    registration knows which names to alias. The expected outputs use
-    fully model-qualified column references (``orders.amount`` etc.) —
-    pinning the post-refactor invariant that all three dialect-aware
-    builders go through ``_resolve_sql`` for the value column AND for
-    the second column on two-arg stats, matching the standard
-    sum/avg/min/max path.
-    """
-
-    def _measure(
-        self,
-        *,
-        agg: str,
-        agg_kwargs: dict[str, str] | None = None,
-    ) -> EnrichedMeasure:
-        return EnrichedMeasure(
-            name="amount",
-            sql="amount",
-            model_name="orders",
-            alias=f"amount_{agg}",
-            aggregation=agg,
-            agg_kwargs=agg_kwargs or {},
-        )
-
-    # --- stddev_samp -------------------------------------------------------
-
-    @pytest.mark.parametrize(
-        "dialect,expected",
-        [
-            ("postgres", "STDDEV_SAMP(orders.amount)"),
-            ("duckdb", "STDDEV_SAMP(orders.amount)"),
-            ("mysql", "STDDEV_SAMP(orders.amount)"),
-            ("sqlite", "STDDEV_SAMP(orders.amount)"),
-            # T-SQL: STDEV is the T-SQL name for sample standard deviation
-            ("tsql", "STDEV(orders.amount)"),
-        ],
-    )
-    def test_build_stddev_samp(self, dialect: str, expected: str) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        m = self._measure(agg="stddev_samp")
-        sql = gen._build_agg(measure=m)[0].sql(dialect=dialect)
-        assert sql == expected
-
-    # --- stddev_pop --------------------------------------------------------
-
-    @pytest.mark.parametrize(
-        "dialect,expected",
-        [
-            ("postgres", "STDDEV_POP(orders.amount)"),
-            ("duckdb", "STDDEV_POP(orders.amount)"),
-            ("mysql", "STDDEV_POP(orders.amount)"),
-            ("sqlite", "STDDEV_POP(orders.amount)"),
-            # T-SQL: STDEVP is the T-SQL name for population standard deviation
-            ("tsql", "STDEVP(orders.amount)"),
-        ],
-    )
-    def test_build_stddev_pop(self, dialect: str, expected: str) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        m = self._measure(agg="stddev_pop")
-        sql = gen._build_agg(measure=m)[0].sql(dialect=dialect)
-        assert sql == expected
-
-    # --- var_samp ----------------------------------------------------------
-
-    @pytest.mark.parametrize(
-        "dialect,expected",
-        [
-            ("postgres", "VAR_SAMP(orders.amount)"),
-            # sqlglot rewrites VAR_SAMP → VARIANCE on SQLite/DuckDB; the
-            # SQLite UDF is therefore registered under the alias `variance`
-            # so generator output still resolves at runtime. MySQL is the
-            # exception: sqlglot's MySQL dialect rewrites the same way, but
-            # MySQL's ``VARIANCE`` is an alias for ``VAR_POP`` (population
-            # variance), so the rewritten SQL would silently return the
-            # wrong value. The generator emits ``VAR_SAMP`` directly on
-            # MySQL via ``exp.Anonymous`` to bypass the transpile.
-            ("duckdb", "VARIANCE(orders.amount)"),
-            ("mysql", "VAR_SAMP(orders.amount)"),
-            ("sqlite", "VARIANCE(orders.amount)"),
-            # T-SQL: VAR is the T-SQL name for sample variance
-            ("tsql", "VAR(orders.amount)"),
-        ],
-    )
-    def test_build_var_samp(self, dialect: str, expected: str) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        m = self._measure(agg="var_samp")
-        sql = gen._build_agg(measure=m)[0].sql(dialect=dialect)
-        assert sql == expected
-
-    # --- var_pop -----------------------------------------------------------
-
-    @pytest.mark.parametrize(
-        "dialect,expected",
-        [
-            ("postgres", "VAR_POP(orders.amount)"),
-            ("duckdb", "VAR_POP(orders.amount)"),
-            # sqlglot rewrites VAR_POP → VARIANCE_POP on SQLite (handled by
-            # a registered UDF alias). MySQL gets the same buggy rewrite,
-            # but ``VARIANCE_POP`` is not a real MySQL function — the
-            # generator emits ``VAR_POP`` directly via ``exp.Anonymous``.
-            ("mysql", "VAR_POP(orders.amount)"),
-            ("sqlite", "VARIANCE_POP(orders.amount)"),
-            # T-SQL: VARP is the T-SQL name for population variance
-            ("tsql", "VARP(orders.amount)"),
-        ],
-    )
-    def test_build_var_pop(self, dialect: str, expected: str) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        m = self._measure(agg="var_pop")
-        sql = gen._build_agg(measure=m)[0].sql(dialect=dialect)
-        assert sql == expected
-
-    # --- corr (2-arg via `other=` kwarg) ----------------------------------
-
-    # corr / covar_samp / covar_pop all share the 2-arg shape and the
-    # `other=` kwarg parameter; parametrize once instead of repeating.
-    @pytest.mark.parametrize(
-        "agg,sql_fn",
-        [
-            ("corr", "CORR"),
-            ("covar_samp", "COVAR_SAMP"),
-            ("covar_pop", "COVAR_POP"),
-        ],
-    )
-    @pytest.mark.parametrize("dialect", ["postgres", "duckdb", "sqlite"])
-    def test_build_two_arg_stat_emits_two_arg_call(
-        self, dialect: str, agg: str, sql_fn: str,
-    ) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        m = self._measure(agg=agg, agg_kwargs={"other": "quantity"})
-        sql = gen._build_agg(measure=m)[0].sql(dialect=dialect)
-        # Both legs go through _resolve_sql, so a bare `quantity` kwarg
-        # qualifies under the LHS measure's model_name.
-        assert sql == f"{sql_fn}(orders.amount, orders.quantity)"
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_clickhouse(self, agg: str) -> None:
-        gen = SQLGenerator(dialect="clickhouse")
-        m = self._measure(agg=agg, agg_kwargs={"other": "quantity"})
-        sql = gen._build_agg(measure=m)[0].sql(dialect="clickhouse")
-        # ClickHouse casing is its own thing; assert the call shape only.
-        assert sql.lower() == f"{agg.lower()}(orders.amount, orders.quantity)"
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_mysql_emits_formula(self, agg: str) -> None:
-        # MySQL has no native CORR / COVAR_SAMP / COVAR_POP but can express them
-        # via the variance-decomposition formula: cov(x,y) = (var(x+y)-var(x)-var(y))/2
-        gen = SQLGenerator(dialect="mysql")
-        m = self._measure(agg=agg, agg_kwargs={"other": "quantity"})
-        sql = gen._build_agg(measure=m)[0].sql(dialect="mysql")
-        # Formula uses MySQL-compatible VAR_SAMP or VAR_POP (covar_pop uses population variance)
-        assert "VAR_SAMP(" in sql or "VAR_POP(" in sql
-        # Not a direct two-arg COVAR_SAMP/COVAR_POP/CORR call (those don't exist in MySQL)
-        assert f"{agg.upper()}(" not in sql
-        # Variance-decomposition uses division
-        assert "/" in sql
-        # Both columns are NULL-guarded against each other
-        assert "CASE WHEN" in sql.upper()
-        # MySQL may emit "IS NOT NULL" or "NOT ... IS NULL" (semantically equivalent)
-        assert "IS NOT NULL" in sql.upper() or "IS NULL" in sql.upper()
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_tsql_emits_formula(self, agg: str) -> None:
-        # T-SQL has no native CORR / COVAR_SAMP / COVAR_POP; use variance-decomposition
-        gen = SQLGenerator(dialect="tsql")
-        m = self._measure(agg=agg, agg_kwargs={"other": "quantity"})
-        sql = gen._build_agg(measure=m)[0].sql(dialect="tsql")
-        # Formula uses T-SQL VAR() or VARP() (covar_pop uses population variance)
-        assert "VAR(" in sql or "VARP(" in sql
-        # Not a direct two-arg call
-        assert f"{agg.upper()}(" not in sql
-        # Variance-decomposition uses division
-        assert "/" in sql
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_mysql_missing_other_prioritises_param_error(
-        self, agg: str,
-    ) -> None:
-        """When BOTH conditions hold (MySQL dialect AND missing `other=`
-        kwarg), the missing-required-param error is more useful to the user
-        than "MySQL not supported" — it points at the actual mistake. Codex
-        review #5 on PR #82: the MySQL guard ran before `other=` resolution.
-        """
-        gen = SQLGenerator(dialect="mysql")
-        m = self._measure(agg=agg, agg_kwargs={})
-        with pytest.raises(ValueError, match=r"requires parameter 'other'"):
-            gen._build_agg(measure=m)
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_missing_other_raises(self, agg: str) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = self._measure(agg=agg, agg_kwargs={})
-        with pytest.raises(ValueError, match=r"requires parameter 'other'|other="):
-            gen._build_agg(measure=m)
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_unsafe_other_rejected(self, agg: str) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = self._measure(
-            agg=agg,
-            agg_kwargs={"other": "quantity); DROP TABLE x; --"},
-        )
-        with pytest.raises(ValueError, match="Unsafe value"):
-            gen._build_agg(measure=m)
-
-    # --- filter wrapping ---------------------------------------------------
-
-    def test_build_stddev_samp_with_filter_wraps_value(self) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = EnrichedMeasure(
-            name="amount",
-            sql="amount",
-            model_name="orders",
-            alias="amount_stddev_samp",
-            aggregation="stddev_samp",
-            agg_kwargs={},
-            filter_sql="status = 'completed'",
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="postgres")
-        # Filter wraps the qualified column reference.
-        assert "CASE WHEN status = 'completed' THEN orders.amount END" in sql
-        assert "STDDEV_SAMP" in sql
-
-    def test_build_corr_with_filter_wraps_both_columns(self) -> None:
-        gen = SQLGenerator(dialect="postgres")
-        m = EnrichedMeasure(
-            name="amount",
-            sql="amount",
-            model_name="orders",
-            alias="amount_corr",
-            aggregation="corr",
-            agg_kwargs={"other": "quantity"},
-            filter_sql="status = 'completed'",
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="postgres")
-        # Both legs of corr() must be wrapped in CASE WHEN so non-matching
-        # rows contribute NULL pairs (which the aggregate skips entirely).
-        assert sql.count("CASE WHEN status = 'completed'") == 2
-        assert "CORR(" in sql
-        # Both legs are also qualified.
-        assert "orders.amount" in sql
-        assert "orders.quantity" in sql
 
 
 class TestStatAggsViaQueryEnrichment:
@@ -6250,213 +5268,6 @@ class TestGetColumnTypesSql:
             f"Empty allowed_aggregations must skip probe, got {formulas}"
         )
 
-
-# DEV-1337: per-alias allowlists for "natively supports single-arg log10/log2"
-# rendering. Outside these sets the current 2-arg LOG(base, x) form is kept.
-# Mirrored in slayer/sql/generator.py — keep in sync.
-_LOG10_NATIVE_DIALECTS = frozenset({
-    "sqlite", "postgres", "duckdb", "mysql", "clickhouse",
-    "snowflake", "bigquery", "redshift",
-    "trino", "presto", "databricks", "spark", "tsql",
-})
-_LOG2_NATIVE_DIALECTS = frozenset({
-    "sqlite", "postgres", "duckdb", "mysql", "clickhouse",
-    "bigquery", "trino", "presto", "databricks", "spark",
-})
-
-
-class TestLogAliasPreservation:
-    """DEV-1337 — user-written ``log10(x)`` / ``log2(x)`` must round-trip
-    verbatim in emitted SQL on dialects that natively support those single-arg
-    aliases. sqlglot's default behaviour normalises both into a generic
-    ``Log(this=Literal(base), expression=arg)`` AST node and re-emits as
-    ``LOG(base, x)`` for almost every dialect, which makes generated SQL
-    diverge from the recipe formula text and (on dialects that lack 2-arg
-    ``LOG``) can break a previously working call.
-    """
-
-    @pytest.fixture
-    def log_model(self) -> SlayerModel:
-        return SlayerModel(
-            name="orders",
-            sql_table="public.orders",
-            data_source="test",
-            columns=[
-                Column(name="id", sql="id", type=DataType.DOUBLE, primary_key=True),
-                Column(name="status", sql="status", type=DataType.TEXT),
-                Column(name="amount", sql="amount", type=DataType.DOUBLE),
-                # Scalar log expressions inside Column.sql — the primary
-                # path the issue surfaced through.
-                Column(name="log_amount", sql="log10(amount)", type=DataType.DOUBLE),
-                Column(name="log2_amount", sql="log2(amount)", type=DataType.DOUBLE),
-                # Negative-control: a non-alias literal base. Must keep the
-                # standard 2-arg LOG(base, x) form post-fix.
-                Column(name="log3_amount", sql="log(3, amount)", type=DataType.DOUBLE),
-                # ln(...) is a separate AST node (exp.Ln); the rewrite must
-                # not touch it.
-                Column(name="ln_amount", sql="ln(amount)", type=DataType.DOUBLE),
-            ],
-        )
-
-    @pytest.mark.parametrize("dialect", TestMultiDialectGeneration.ALL_DIALECTS)
-    async def test_log10_in_column_sql_is_preserved(
-        self, dialect: str, log_model: SlayerModel,
-    ) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="log_amount:max")],
-        )
-        sql = await _generate(generator=gen, query=query, model=log_model)
-        upper_no_ws = "".join(sql.upper().split())
-        if dialect in _LOG10_NATIVE_DIALECTS:
-            assert "LOG10(AMOUNT)" in upper_no_ws, (
-                f"{dialect}: expected literal LOG10(amount), got:\n{sql}"
-            )
-            # Must not have canonicalised to either arg-order 2-arg form.
-            assert "LOG(10,AMOUNT)" not in upper_no_ws, (
-                f"{dialect}: should not canonicalise to LOG(10, amount):\n{sql}"
-            )
-            assert "LOG(AMOUNT,10)" not in upper_no_ws, (
-                f"{dialect}: should not canonicalise to LOG(amount, 10):\n{sql}"
-            )
-        else:
-            # Fallback: current 2-arg LOG behaviour is preserved on dialects
-            # without native single-arg log10 (oracle).
-            assert "LOG(10,AMOUNT)" in upper_no_ws or "LOG(AMOUNT,10)" in upper_no_ws, (
-                f"{dialect}: expected fallback LOG(base,x) form, got:\n{sql}"
-            )
-
-    @pytest.mark.parametrize("dialect", TestMultiDialectGeneration.ALL_DIALECTS)
-    async def test_log2_in_column_sql_is_preserved(
-        self, dialect: str, log_model: SlayerModel,
-    ) -> None:
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="log2_amount:max")],
-        )
-        sql = await _generate(generator=gen, query=query, model=log_model)
-        upper_no_ws = "".join(sql.upper().split())
-        if dialect in _LOG2_NATIVE_DIALECTS:
-            assert "LOG2(AMOUNT)" in upper_no_ws, (
-                f"{dialect}: expected literal LOG2(amount), got:\n{sql}"
-            )
-            assert "LOG(2,AMOUNT)" not in upper_no_ws, (
-                f"{dialect}: should not canonicalise to LOG(2, amount):\n{sql}"
-            )
-            assert "LOG(AMOUNT,2)" not in upper_no_ws, (
-                f"{dialect}: should not canonicalise to LOG(amount, 2):\n{sql}"
-            )
-        else:
-            # Fallback for tsql / oracle / redshift / snowflake (no native LOG2).
-            assert "LOG(2,AMOUNT)" in upper_no_ws or "LOG(AMOUNT,2)" in upper_no_ws, (
-                f"{dialect}: expected fallback LOG(base,x) form, got:\n{sql}"
-            )
-
-    @pytest.mark.parametrize("dialect", sorted(_LOG10_NATIVE_DIALECTS))
-    async def test_log10_inside_filtered_column_survives_reparse(
-        self, dialect: str,
-    ) -> None:
-        """A filtered Column (``Column.filter="..."``) wraps the resolved
-        value in ``CASE WHEN ... THEN ... END`` and re-parses through
-        sqlglot. The log-alias rewrite must survive that round-trip — a
-        re-parse of ``LOG10(amount)`` would otherwise canonicalise back to
-        a generic ``Log`` node and re-emit as ``LOG(10, amount)``.
-        """
-        model = SlayerModel(
-            name="orders",
-            sql_table="public.orders",
-            data_source="test",
-            columns=[
-                Column(name="id", sql="id", type=DataType.DOUBLE, primary_key=True),
-                Column(name="status", sql="status", type=DataType.TEXT),
-                Column(name="amount", sql="amount", type=DataType.DOUBLE),
-                Column(
-                    name="log_completed_amount",
-                    sql="log10(amount)",
-                    filter="status = 'completed'",
-                    type=DataType.DOUBLE,
-                ),
-            ],
-        )
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="log_completed_amount:max")],
-        )
-        sql = await _generate(generator=gen, query=query, model=model)
-        upper_no_ws = "".join(sql.upper().split())
-        assert "LOG10(AMOUNT)" in upper_no_ws, (
-            f"{dialect}: expected literal log10(amount) inside filtered "
-            f"column wrapper, got:\n{sql}"
-        )
-        assert "LOG(10,AMOUNT)" not in upper_no_ws, (
-            f"{dialect}: filtered-column re-parse must not re-canonicalise "
-            f"to LOG(10, amount):\n{sql}"
-        )
-
-    @pytest.mark.parametrize("dialect", sorted(_LOG10_NATIVE_DIALECTS))
-    async def test_log10_in_arithmetic_measure_is_preserved(
-        self, dialect: str, log_model: SlayerModel,
-    ) -> None:
-        """Arithmetic that mixes a log10 column-derived measure with COUNT(*).
-        Pins that the rewrite survives the arithmetic enrichment path.
-        """
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="log_amount:max / *:count", name="ratio")],
-        )
-        sql = await _generate(generator=gen, query=query, model=log_model)
-        upper_no_ws = "".join(sql.upper().split())
-        assert "LOG10(AMOUNT)" in upper_no_ws, (
-            f"{dialect}: expected log10(amount) inside arithmetic measure:\n{sql}"
-        )
-        assert "COUNT(" in sql.upper(), f"COUNT(*) leg missing on {dialect}:\n{sql}"
-
-    @pytest.mark.parametrize("dialect", TestMultiDialectGeneration.ALL_DIALECTS)
-    async def test_log_with_non_alias_base_unchanged(
-        self, dialect: str, log_model: SlayerModel,
-    ) -> None:
-        """Negative test: ``log(3, amount)`` (literal base ≠ 10/2) must keep
-        the standard 2-arg form. The rewrite is scoped to bases 10 and 2 only.
-        """
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="log3_amount:max")],
-        )
-        sql = await _generate(generator=gen, query=query, model=log_model)
-        upper_no_ws = "".join(sql.upper().split())
-        # Must NOT have invented a single-arg LOG3(...) function.
-        assert "LOG3(" not in upper_no_ws, (
-            f"{dialect}: must not invent LOG3() — only base 10 and 2 are aliased:\n{sql}"
-        )
-        # The 2-arg form must remain in some arg order.
-        assert "LOG(3,AMOUNT)" in upper_no_ws or "LOG(AMOUNT,3)" in upper_no_ws, (
-            f"{dialect}: expected 2-arg LOG(3, amount) preserved, got:\n{sql}"
-        )
-
-    @pytest.mark.parametrize("dialect", TestMultiDialectGeneration.ALL_DIALECTS)
-    async def test_ln_unchanged(
-        self, dialect: str, log_model: SlayerModel,
-    ) -> None:
-        """``ln(x)`` lives under a separate sqlglot AST node (``exp.Ln``);
-        the rewrite must not affect it.
-        """
-        gen = SQLGenerator(dialect=dialect)
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="ln_amount:max")],
-        )
-        sql = await _generate(generator=gen, query=query, model=log_model)
-        # T-SQL has no LN — sqlglot transpiles to LOG(x). Every other dialect
-        # keeps LN(...). We only assert the rewrite did not invent something.
-        assert "LN10(" not in sql.upper()
-        assert "LN2(" not in sql.upper()
-
-
 # ---------------------------------------------------------------------------
 # DEV-1336 — window functions in filters (single-stage)
 # ---------------------------------------------------------------------------
@@ -6871,115 +5682,6 @@ class TestCastEmissionNonBasePaths:
         # CAST present somewhere — column_type propagated to formula expansion.
         assert "CAST(" in sql.upper()
         assert "JSON_EXTRACT" in sql.upper()
-
-
-class TestStringHygieneDialectTranslation:
-    """DEV-1378: lowercase string-hygiene operators are pass-through to
-    the emitted SQL string, then re-parsed by sqlglot under the target
-    dialect at WHERE-assembly. sqlglot's per-dialect emitter chooses
-    each dialect's preferred spelling. These tests pin the actual
-    emitted SQL across SQLite / Postgres / MySQL / DuckDB / ClickHouse
-    so a future sqlglot upgrade that changes the spelling is caught.
-    """
-
-    @pytest.mark.parametrize(
-        "dialect,expected",
-        [
-            ("sqlite", "LOWER(orders.status) = 'active'"),
-            ("postgres", "LOWER(orders.status) = 'active'"),
-            ("mysql", "LOWER(orders.status) = 'active'"),
-            ("duckdb", "LOWER(orders.status) = 'active'"),
-            ("clickhouse", "lower(orders.status) = 'active'"),
-        ],
-    )
-    async def test_lower(self, orders_model: SlayerModel, dialect: str, expected: str) -> None:
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="*:count")],
-            filters=["lower(status) = 'active'"],
-        )
-        sql = await _generate(
-            generator=SQLGenerator(dialect=dialect),
-            query=query,
-            model=orders_model,
-        )
-        assert expected in sql, f"{dialect}: {expected!r} not in {sql!r}"
-
-    @pytest.mark.parametrize(
-        "dialect,expected",
-        [
-            ("sqlite", "INSTR(orders.status, ',')"),
-            ("postgres", "POSITION(',' IN orders.status)"),
-            ("mysql", "LOCATE(',', orders.status)"),
-            ("duckdb", "STRPOS(orders.status, ',')"),
-            ("clickhouse", "POSITION(orders.status, ',')"),
-        ],
-    )
-    async def test_instr_translates_per_dialect(
-        self, orders_model: SlayerModel, dialect: str, expected: str,
-    ) -> None:
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="*:count")],
-            filters=["instr(status, ',') > 0"],
-        )
-        sql = await _generate(
-            generator=SQLGenerator(dialect=dialect),
-            query=query,
-            model=orders_model,
-        )
-        assert expected in sql, f"{dialect}: {expected!r} not in {sql!r}"
-
-    @pytest.mark.parametrize(
-        "dialect,expected",
-        [
-            ("sqlite", "SUBSTRING(orders.status, 1, 5)"),
-            ("postgres", "SUBSTRING(orders.status FROM 1 FOR 5)"),
-            ("mysql", "SUBSTRING(orders.status, 1, 5)"),
-            ("duckdb", "SUBSTRING(orders.status, 1, 5)"),
-            ("clickhouse", "substr(orders.status, 1, 5)"),
-        ],
-    )
-    async def test_substr_translates_per_dialect(
-        self, orders_model: SlayerModel, dialect: str, expected: str,
-    ) -> None:
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="*:count")],
-            filters=["substr(status, 1, 5) = 'abcde'"],
-        )
-        sql = await _generate(
-            generator=SQLGenerator(dialect=dialect),
-            query=query,
-            model=orders_model,
-        )
-        assert expected in sql, f"{dialect}: {expected!r} not in {sql!r}"
-
-    @pytest.mark.parametrize(
-        "dialect,expected_substring",
-        [
-            # SQLite normalises CONCAT(...) → a || b at emit time.
-            ("sqlite", "orders.status || orders.status"),
-            ("postgres", "CONCAT(orders.status, orders.status)"),
-            ("mysql", "CONCAT(orders.status, orders.status)"),
-            ("duckdb", "CONCAT(orders.status, orders.status)"),
-            ("clickhouse", "CONCAT(orders.status, orders.status)"),
-        ],
-    )
-    async def test_concat_via_pipe_pipe_translates_per_dialect(
-        self, orders_model: SlayerModel, dialect: str, expected_substring: str,
-    ) -> None:
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="*:count")],
-            filters=["status || status = 'foo'"],
-        )
-        sql = await _generate(
-            generator=SQLGenerator(dialect=dialect),
-            query=query,
-            model=orders_model,
-        )
-        assert expected_substring in sql, f"{dialect}: {expected_substring!r} not in {sql!r}"
 
 
 class TestReplaceFunctionInPredicate:
@@ -7856,292 +6558,3 @@ class TestFilterOuterParenWrapDev1539:
         assert "SUM(amount)" in result
 
 
-class TestTsqlDialect:
-    """DEV-1520: T-SQL (SQL Server) dialect-specific SQL generation tests."""
-
-    @pytest.fixture
-    def gen(self) -> SQLGenerator:
-        return SQLGenerator(dialect="tsql")
-
-    @pytest.fixture
-    def orders_model(self) -> SlayerModel:
-        return SlayerModel(
-            name="orders",
-            sql_table="dbo.orders",
-            data_source="test",
-            default_time_dimension="created_at",
-            columns=[
-                Column(name="id", sql="id", type=DataType.INT, primary_key=True),
-                Column(name="status", sql="status", type=DataType.TEXT),
-                Column(name="created_at", sql="created_at", type=DataType.TIMESTAMP),
-                Column(name="revenue", sql="amount", type=DataType.DOUBLE),
-                Column(name="quantity", sql="quantity", type=DataType.DOUBLE),
-            ],
-        )
-
-    # --- date trunc ---
-
-    def test_build_date_trunc_month_emits_datetrunc(self, gen: SQLGenerator) -> None:
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_date_trunc(col, TimeGranularity.MONTH).sql(dialect="tsql")
-        assert "DATETRUNC" in sql.upper()
-        assert "MONTH" in sql.upper()
-
-    def test_build_date_trunc_year(self, gen: SQLGenerator) -> None:
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_date_trunc(col, TimeGranularity.YEAR).sql(dialect="tsql")
-        assert "DATETRUNC" in sql.upper()
-        assert "YEAR" in sql.upper()
-
-    def test_build_date_trunc_day(self, gen: SQLGenerator) -> None:
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_date_trunc(col, TimeGranularity.DAY).sql(dialect="tsql")
-        assert "DATETRUNC" in sql.upper()
-        assert "DAY" in sql.upper()
-
-    def test_build_date_trunc_week_uses_iso_week(self, gen: SQLGenerator) -> None:
-        """Week truncation must use ISO_WEEK for Monday-start (@@DATEFIRST-independent)."""
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_date_trunc(col, TimeGranularity.WEEK).sql(dialect="tsql")
-        assert "ISO_WEEK" in sql.upper(), (
-            f"T-SQL week truncation must use ISO_WEEK (not WEEK) to be "
-            f"locale-independent. Got: {sql}"
-        )
-        # DATETRUNC(WEEK, ...) without ISO_ is locale-dependent — must not appear
-        assert "DATETRUNC(WEEK" not in sql.upper().replace("ISO_WEEK", ""), (
-            f"T-SQL week truncation must not use bare WEEK (@@DATEFIRST-dependent): {sql}"
-        )
-
-    def test_build_date_trunc_quarter(self, gen: SQLGenerator) -> None:
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_date_trunc(col, TimeGranularity.QUARTER).sql(dialect="tsql")
-        assert "DATETRUNC" in sql.upper()
-        assert "QUARTER" in sql.upper()
-
-    def test_build_date_trunc_no_date_trunc_function(self, gen: SQLGenerator) -> None:
-        """T-SQL uses DATETRUNC (no underscore), not DATE_TRUNC."""
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_date_trunc(col, TimeGranularity.MONTH).sql(dialect="tsql")
-        assert "DATE_TRUNC" not in sql.upper(), f"T-SQL should use DATETRUNC, got: {sql}"
-
-    # --- time offset ---
-
-    def test_build_time_offset_year(self, gen: SQLGenerator) -> None:
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_time_offset_expr(col, -1, "year").sql(dialect="tsql")
-        assert "DATEADD" in sql.upper()
-        assert "YEAR" in sql.upper()
-
-    def test_build_time_offset_month(self, gen: SQLGenerator) -> None:
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_time_offset_expr(col, -1, "month").sql(dialect="tsql")
-        assert "DATEADD" in sql.upper()
-        assert "MONTH" in sql.upper()
-
-    def test_build_time_offset_positive(self, gen: SQLGenerator) -> None:
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_time_offset_expr(col, 3, "day").sql(dialect="tsql")
-        assert "DATEADD" in sql.upper()
-        assert "DAY" in sql.upper()
-        assert "3" in sql
-        assert "created_at" in sql
-        assert "INTERVAL" not in sql.upper()
-
-    @pytest.mark.parametrize("gran", ["year", "month", "day", "week"])
-    def test_build_time_offset_no_interval_keyword(self, gen: SQLGenerator, gran: str) -> None:
-        """T-SQL must never emit INTERVAL (invalid syntax) for time offsets."""
-        col = sqlglot.parse_one("created_at", dialect="tsql")
-        sql = gen._build_time_offset_expr(col, -1, gran).sql(dialect="tsql")
-        assert "INTERVAL" not in sql.upper(), (
-            f"INTERVAL is invalid T-SQL syntax for granularity {gran!r}: {sql}"
-        )
-
-    # --- median / percentile (unsupported) ---
-
-    def test_build_median_tsql_raises(self, gen: SQLGenerator) -> None:
-        """T-SQL PERCENTILE_CONT is window-only (requires OVER); unsupported as GROUP BY agg."""
-        inner = sqlglot.parse_one("amount", dialect="tsql")
-        with pytest.raises(NotImplementedError):
-            gen._build_median(inner)
-
-    def test_build_percentile_tsql_raises(self, gen: SQLGenerator) -> None:
-        """T-SQL PERCENTILE_CONT is window-only (requires OVER); unsupported as GROUP BY agg."""
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias="amount_percentile", aggregation="percentile",
-            agg_kwargs={"p": "0.5"},
-        )
-        with pytest.raises(NotImplementedError):
-            gen._build_percentile(m)
-
-    # --- one-arg stat aggs ---
-
-    @pytest.mark.parametrize("agg,expected_fn", [
-        ("stddev_samp", "STDEV"),
-        ("stddev_pop", "STDEVP"),
-        ("var_samp", "VAR"),
-        ("var_pop", "VARP"),
-    ])
-    def test_build_one_arg_stat_tsql(
-        self, gen: SQLGenerator, agg: str, expected_fn: str,
-    ) -> None:
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias=f"amount_{agg}", aggregation=agg, agg_kwargs={},
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="tsql")
-        assert f"{expected_fn}(" in sql, f"Expected {expected_fn}() in {sql!r}"
-        assert "orders.amount" in sql
-
-    # --- two-arg stat aggs (variance-decomposition formula) ---
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_tsql_uses_var_function(
-        self, gen: SQLGenerator, agg: str,
-    ) -> None:
-        """covar/corr on T-SQL must use T-SQL VAR() not Postgres VAR_SAMP()."""
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias=f"amount_{agg}", aggregation=agg,
-            agg_kwargs={"other": "quantity"},
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="tsql")
-        # covar_samp/corr use VAR(), covar_pop uses VARP() — both are valid T-SQL
-        assert "VAR(" in sql or "VARP(" in sql, f"Expected VAR()/VARP() in formula, got: {sql}"
-        # Must NOT use Postgres-style VAR_SAMP (invalid T-SQL function)
-        assert "VAR_SAMP(" not in sql, f"VAR_SAMP is not a T-SQL function: {sql}"
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_tsql_no_direct_call(
-        self, gen: SQLGenerator, agg: str,
-    ) -> None:
-        """T-SQL doesn't have COVAR_SAMP / COVAR_POP / CORR natively."""
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias=f"amount_{agg}", aggregation=agg,
-            agg_kwargs={"other": "quantity"},
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="tsql")
-        assert f"{agg.upper()}(" not in sql, (
-            f"T-SQL should not emit a direct {agg.upper()}() call; use formula. Got: {sql}"
-        )
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_tsql_contains_both_columns(
-        self, gen: SQLGenerator, agg: str,
-    ) -> None:
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias=f"amount_{agg}", aggregation=agg,
-            agg_kwargs={"other": "quantity"},
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="tsql")
-        assert "orders.amount" in sql
-        assert "orders.quantity" in sql
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_tsql_uses_division(
-        self, gen: SQLGenerator, agg: str,
-    ) -> None:
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias=f"amount_{agg}", aggregation=agg,
-            agg_kwargs={"other": "quantity"},
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="tsql")
-        assert "/" in sql, f"Variance-decomposition formula must contain division: {sql}"
-
-    @pytest.mark.parametrize("agg", ["covar_samp", "covar_pop"])
-    def test_build_covar_tsql_null_guards_both_columns(
-        self, gen: SQLGenerator, agg: str,
-    ) -> None:
-        """Both columns must be NULL-guarded against each other in the formula.
-
-        For `covar_samp(x, y)`: x is guarded as `CASE WHEN y IS NOT NULL THEN x END`
-        and y is guarded as `CASE WHEN x IS NOT NULL THEN y END`, so pairs where
-        either column is NULL are excluded from the variance computation.
-        """
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias=f"amount_{agg}", aggregation=agg,
-            agg_kwargs={"other": "quantity"},
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="tsql")
-        upper = sql.upper()
-        # Both x-guarded-by-y and y-guarded-by-x CASE WHEN patterns must appear
-        assert "CASE WHEN" in upper, f"Expected NULL guards (CASE WHEN) in formula: {sql}"
-        # T-SQL may emit "IS NOT NULL" or "NOT ... IS NULL" (semantically equivalent)
-        assert "IS NOT NULL" in upper or "IS NULL" in upper, (
-            f"Expected IS NULL/IS NOT NULL guard in formula: {sql}"
-        )
-
-    def test_build_corr_tsql_uses_stdev_for_denominator(self, gen: SQLGenerator) -> None:
-        """corr denominator uses STDEV (T-SQL stddev_samp) * STDEV."""
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias="amount_corr", aggregation="corr",
-            agg_kwargs={"other": "quantity"},
-        )
-        sql = gen._build_agg(measure=m)[0].sql(dialect="tsql")
-        assert "STDEV(" in sql, f"Expected STDEV() in corr denominator, got: {sql}"
-
-    @pytest.mark.parametrize("agg", ["corr", "covar_samp", "covar_pop"])
-    def test_build_two_arg_stat_tsql_missing_other_raises(
-        self, gen: SQLGenerator, agg: str,
-    ) -> None:
-        m = EnrichedMeasure(
-            name="amount", sql="amount", model_name="orders",
-            alias=f"amount_{agg}", aggregation=agg, agg_kwargs={},
-        )
-        with pytest.raises(ValueError, match=r"requires parameter 'other'|other="):
-            gen._build_agg(measure=m)
-
-    # --- full query integration ---
-
-    async def test_full_aggregation_query_valid_tsql(
-        self, gen: SQLGenerator, orders_model: SlayerModel,
-    ) -> None:
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="*:count"), ModelMeasure(formula="revenue:sum")],
-            dimensions=[ColumnRef(name="status")],
-        )
-        sql = await _generate(gen, query, orders_model)
-        assert "COUNT(" in sql
-        assert "SUM(" in sql
-
-    async def test_full_query_with_time_dim_valid_tsql(
-        self, gen: SQLGenerator, orders_model: SlayerModel,
-    ) -> None:
-        query = SlayerQuery(
-            source_model="orders",
-            measures=[ModelMeasure(formula="revenue:sum")],
-            time_dimensions=[TimeDimension(
-                dimension=ColumnRef(name="created_at"),
-                granularity=TimeGranularity.MONTH,
-            )],
-        )
-        sql = await _generate(gen, query, orders_model)
-        assert "DATETRUNC" in sql.upper()
-        assert "SUM(" in sql
-
-    async def test_calendar_time_shift_tsql_uses_dateadd(
-        self, gen: SQLGenerator, orders_model: SlayerModel,
-    ) -> None:
-        query = SlayerQuery(
-            source_model="orders",
-            time_dimensions=[TimeDimension(
-                dimension=ColumnRef(name="created_at"),
-                granularity=TimeGranularity.MONTH,
-            )],
-            measures=[
-                ModelMeasure(formula="revenue:sum"),
-                ModelMeasure(formula="time_shift(revenue:sum, -1, 'year')", name="rev_prev_year"),
-            ],
-        )
-        sql = await _generate(gen, query, orders_model)
-        assert "shifted_" in sql
-        assert "DATEADD" in sql.upper()
-        assert "INTERVAL" not in sql.upper(), (
-            f"INTERVAL is invalid T-SQL syntax; shifted CTE must use DATEADD:\n{sql}"
-        )
